Move tests that only apply to multistore_file out of test_file.py
